requirements.txt
epubcheck~=4.2.6
python-slugify~=8.0.1
vBulletinThreadUtils~=0.1
EbookLib~=0.18